Make not-found-route consistent across client and server
