+++ src/server/routing/prepare-app-routes.js
@@ -1,23 +1,19 @@
-import defaultRoutes from './default-routes'
 import buildErrorView from '../render/error-view'
 
 export default config => {
-  // if user routes have been defined, take those in preference to the defaults
-  const routes = config.routes || defaultRoutes(config.components)
-  let preparedRoutes = routes.map(route => {
-    return {
-      ...route,
-      exact: true
-    }
-  })
-  // Add the a default 404 route
+  let routes = config.routes
+  // import and use default routes if user has not defined any routes
+  if (!routes) routes = require('./default-routes').default(config.components)
+  // each route needs to be unambiguous
+  const preparedRoutes = routes.map(route => ({ ...route, exact: true }))
+  // add the a default 404 route
   preparedRoutes.push({
+    path: '(.*)',
     notFoundRoute: true,
     component: buildErrorView({
       config,
       missing: false
     })
   })
-
   return preparedRoutes
 }
